move fake_clip_model fixture
